requirements.txt
pymysql
setuptools
sqlalchemy
python-decouple